requirements.txt
fastapi==0.115.14
openai==1.93.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlalchemy==2.0.41
pgvector==0.4.1
uvicorn==0.35.0
langchain-text-splitters==0.3.8